requirements.txt
flask==3.1.1
python-dotenv==1.0.1
google-generativeai==0.8.5
flask-cors==6.0.1
